Split Sulfide, SulfideElement and Selectors in separate modules
This commits moves the code for SulfideElement and for the Sulfide
selectors to their own files.
